+++ Mainframe3270/__init__.py
@@ -9,7 +9,8 @@ from robot.utils import ConnectionCache
 from robotlibcore import DynamicCore
 
 from Mainframe3270.keywords import (AssertionKeywords, ConnectionKeywords,
-                                    ReadWriteKeywords, WaitAndTimeoutKeywords)
+                                    ReadWriteKeywords, ScreenshotKeywords,
+                                    WaitAndTimeoutKeywords)
 from Mainframe3270.py3270 import Emulator
 from Mainframe3270.utils import convert_timeout
 from Mainframe3270.version import VERSION
@@ -117,8 +118,9 @@ class Mainframe3270(DynamicCore):
         libraries = [
             AssertionKeywords(self),
             ConnectionKeywords(self),
-            WaitAndTimeoutKeywords(self),
             ReadWriteKeywords(self),
+            ScreenshotKeywords(self),
+            WaitAndTimeoutKeywords(self),
             X3270(self),
         ]
         DynamicCore.__init__(self, libraries)
