+++ utest/Mainframe3270/keywords/test_screenshot.py
@@ -1,12 +1,20 @@
 import os
 
+import pytest
 from pytest_mock import MockerFixture
 from robot.api import logger
 
-from Mainframe3270.x3270 import X3270
+from Mainframe3270.keywords import ScreenshotKeywords
+
+from .utils import create_test_object_for
 
 
-def test_set_screenshot_folder(under_test: X3270):
+@pytest.fixture
+def under_test():
+    return create_test_object_for(ScreenshotKeywords)
+
+
+def test_set_screenshot_folder(under_test: ScreenshotKeywords):
     path = os.getcwd()
 
     under_test.set_screenshot_folder(path)
@@ -14,7 +22,9 @@ def test_set_screenshot_folder(under_test: X3270):
     assert under_test.img_folder == os.getcwd()
 
 
-def test_set_screenshot_folder_nonexistent(mocker: MockerFixture, under_test: X3270):
+def test_set_screenshot_folder_nonexistent(
+    mocker: MockerFixture, under_test: ScreenshotKeywords
+):
     mocker.patch("robot.api.logger.error")
     mocker.patch("robot.api.logger.warn")
     path = os.path.join(os.getcwd(), "nonexistent")
@@ -27,7 +37,7 @@ def test_set_screenshot_folder_nonexistent(mocker: MockerFixture, under_test: X3
     )
 
 
-def test_take_screenshot(mocker: MockerFixture, under_test: X3270):
+def test_take_screenshot(mocker: MockerFixture, under_test: ScreenshotKeywords):
     mocker.patch("Mainframe3270.py3270.Emulator.save_screen")
     mocker.patch("robot.api.logger.write")
     mocker.patch("time.time", return_value=1.0)
@@ -44,6 +54,13 @@ def test_take_screenshot(mocker: MockerFixture, under_test: X3270):
     else:
         assert filepath == "./screenshot_1000.html"
 
+
+def test_take_screenshot_with_filename_prefix(
+    mocker: MockerFixture, under_test: ScreenshotKeywords
+):
+    mocker.patch("Mainframe3270.py3270.Emulator.save_screen")
+    mocker.patch("robot.api.logger.write")
+    mocker.patch("time.time", return_value=1.0)
     filepath = under_test.take_screenshot(250, 250, "MyScreenshot")
 
     logger.write.assert_called_with(
